Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/svg/painting/parsing/image-rendering-computed-expected.txt


PASS Property image-rendering value 'auto' computes to 'auto' 
FAIL Property image-rendering value 'optimizequality' computes to 'optimizequality' assert_equals: expected "optimizequality" but got "optimizeQuality"
FAIL Property image-rendering value 'optimizespeed' computes to 'optimizespeed' assert_equals: expected "optimizespeed" but got "optimizeSpeed"
